backend/requirements.txt
fastapi
uvicorn[standard]
openai==0.28
python-dotenv